Add bulk upload

+++ django/shrubbery/web/teacher_views.py
@@ -32,6 +32,50 @@ def add_participant(request):
 
 
 @is_teacher
+def add_participants(request):
+    '''Add new participants from a file.'''
+    try:
+        csv_file = request.FILES["csv_file"]
+        file_data = csv_file.read().decode("utf-8")
+        lines = file_data.split("\n")
+    except:
+        context = {
+            'participants': Student.objects.all(),
+            'errors': {'csv': 'Грешка при обработка на файла.'}
+        }
+        return render(request, "participants/participants.html", context)
+    errors = []
+    for i, line in enumerate(lines):
+        if not line:
+            continue
+        try:
+            first_name, last_name, fn = line.split(',')
+            first_name = first_name.strip()
+            last_name = last_name.strip()
+            fn = fn.strip()
+            data = {
+                'first_name': first_name,
+                'last_name': last_name,
+                'fn': fn
+            }
+            form = AddStudentForm(data)
+            if form.is_valid():
+                student = form.save(commit=False)
+                student.email = f'{student.fn}@notset.com'
+                student.is_active = False
+                student.save()
+            else:
+                errors.append((i, line, str(form.errors)))
+        except Exception as e:
+            errors.append((i, line, f'Грешка при обработка на реда: {e}'))
+    context = {
+        'participants': Student.objects.all(),
+        'errors': {'csv_list': errors}
+    }
+    return render(request, "participants/participants.html", context)
+
+
+@is_teacher
 def participant(request, participant):
     '''Edit a single participant.'''
     try:
@@ -48,7 +92,7 @@ def participant(request, participant):
                 'errors': form.errors
             }
             return render(request, "participants/participant.html", context)
-        else:
+        elif 'delete' in request.POST:
             participant_obj.delete()
             return redirect('web:participants')
     else:
